pyproject.toml
[project]
name = "publishing"
version = "0.1.0"
description = "Internal publication system. Create, version, and search authoritative content."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "yt-dlp>=2024.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "psycopg[binary,pool]>=3.1.0",
    "pgvector>=0.2.0",
    "gitpython>=3.1.0",
    "tiktoken>=0.5.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "alembic>=1.17.2",
    "sqlalchemy>=2.0.45",
    "trafilatura>=2.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
pub = "publishing.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/publishing"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
